rename compiled ext to _cutils, explicit import in FFTKDE
