+++ KDEpy/FFTKDE.py
@@ -10,8 +10,10 @@ import numpy as np
 from scipy.signal import convolve
 
 from KDEpy.BaseKDE import BaseKDE
-from KDEpy.binning import grid_is_sorted, linear_binning
+from KDEpy.binning import linear_binning
 from KDEpy.utils import cartesian
+# noinspection PyProtectedMember, PyUnresolvedReferences
+from KDEpy._cutils import grid_is_sorted
 
 
 class FFTKDE(BaseKDE):
